+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 ### Changes this version:
-- Added PWM direction toggle
+- Added effect monitoring per axis
  
 ### Changes in 1.13.x
+- Added PWM direction toggle
 - Added basic iterative TMC PI autotuning
 - Fixed issues with CAN transmission with multiple axes
 - Added SSI encoder support (AMT232B)